--- Doraemon.Data/config.example.json
@@ -1,15 +1,0 @@
-﻿{
-  "MainGuildId": "", // The ID of the guild that the bot will be ran in.
-  "Token":  "", // Your bot's token.
-  "Prefix": "", // The prefix of the bot.
-  "DbConnection": "", // The connection string to your postgresql database.
-  "StaffRoleId": "", // The role that all of your staff members have. Like the general "Staff" role.
-  "PromotionRoleId": "", // The roleId that users can be promoted to via campaigns.
-  "ModmailCategoryId": "", // The category ID that modmail threads should be created in.
-  "LogConfiguration": {
-    "PromotionLogChannelId": "", // Where promotions are logged.
-    "ModLogChannelId": "", // Where moderation actions are logged.
-    "UserJoinedLogChannelId": "",// Where user-joins are logged.
-    "MessageLogChannelId":  ""// Where message edits and deletes.
-  }
-}
