+++ Doraemon/Modules/ModerationModule.cs
@@ -13,7 +13,8 @@ using Doraemon.Common.Extensions;
 using Doraemon.Data;
 using Doraemon.Data.Models;
 using Doraemon.Data.Models.Core;
-using Doraemon.Data.Services;
+using Doraemon.Services.Moderation;
+using Doraemon.Services.Core;
 using Doraemon.Common;
 using Microsoft.EntityFrameworkCore;
 using System.Data;
